python/examples/dat-evals-latent-logging/requirements.txt
arize-phoenix-evals>=0.0.5
openai>=1
arize[Tracing]
arize>=7.1.0
packaging
